backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
PyPDF2==3.0.1

# AI Providers (optional - install based on your choice)
openai==1.10.0
anthropic==0.18.1
google-genai>=1.60.0

# Optional: for testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
